server/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
Flask-Cors==4.0.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
PyMySQL==1.1.0
PyYAML==6.0.1
SQLAlchemy==2.0.29
typing-extensions==4.11.0
urllib3==2.2.1
werkzeug==3.0.2
zipp==3.18.1
gunicorn==22.0.0
python-dotenv==1.0.1
google-cloud-secret-manager==2.20.0